+++ crates/miden-agglayer/src/claim_note.rs
@@ -10,14 +10,13 @@ use miden_protocol::errors::NoteError;
 use miden_protocol::note::{
     Note,
     NoteAssets,
-    NoteExecutionHint,
     NoteMetadata,
     NoteRecipient,
     NoteStorage,
     NoteTag,
     NoteType,
 };
-use miden_standards::note::NetworkAccountTarget;
+use miden_standards::note::{NetworkAccountTarget, NoteExecutionHint};
 
 use crate::utils::bytes32_to_felts;
 use crate::{EthAddressFormat, EthAmount, claim_script};