+++ crates/miden-testing/src/kernel_tests/tx/test_faucet.rs
@@ -36,6 +36,7 @@ use crate::{TransactionContextBuilder, assert_execution_error, assert_transactio
 #[tokio::test]
 async fn test_mint_fungible_asset_succeeds() -> anyhow::Result<()> {
     let faucet_id = AccountId::try_from(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET).unwrap();
+    let asset = FungibleAsset::new(faucet_id, FUNGIBLE_ASSET_AMOUNT)?;
 
     let code = format!(
         r#"
@@ -49,22 +50,28 @@ async fn test_mint_fungible_asset_succeeds() -> anyhow::Result<()> {
             exec.prologue::prepare_transaction
 
             # mint asset
-            push.{FUNGIBLE_ASSET_AMOUNT} push.0 push.{suffix} push.{prefix}
+            push.{FUNGIBLE_ASSET}
             call.mock_faucet::mint
 
             # assert the correct asset is returned
-            push.{FUNGIBLE_ASSET_AMOUNT} push.0 push.{suffix} push.{prefix}
+            push.{FUNGIBLE_ASSET}
             assert_eqw.err="minted asset does not match expected asset"
 
             # assert the input vault has been updated
             exec.memory::get_input_vault_root_ptr
-            push.{suffix} push.{prefix}
-            exec.asset_vault::get_balance
+            push.{ASSET_KEY}
+            exec.asset_vault::get_asset
+            # => [ASSET]
+
+            # extract balance from asset
+            drop drop drop
+            # => [balance]
+
             push.{FUNGIBLE_ASSET_AMOUNT} assert_eq.err="input vault should contain minted asset"
         end
         "#,
-        prefix = faucet_id.prefix().as_felt(),
-        suffix = faucet_id.suffix(),
+        FUNGIBLE_ASSET = Word::from(asset),
+        ASSET_KEY = asset.vault_key(),
     );
 
     TransactionContextBuilder::with_fungible_faucet(faucet_id.into())
@@ -176,7 +183,6 @@ async fn test_mint_non_fungible_asset_succeeds() -> anyhow::Result<()> {
     let tx_context =
         TransactionContextBuilder::with_non_fungible_faucet(NonFungibleAsset::mock_issuer().into())
             .build()?;
-
     let non_fungible_asset = NonFungibleAsset::mock(&NON_FUNGIBLE_ASSET_DATA);
 
     let code = format!(
@@ -201,13 +207,15 @@ async fn test_mint_non_fungible_asset_succeeds() -> anyhow::Result<()> {
 
             # assert the input vault has been updated.
             exec.memory::get_input_vault_root_ptr
+            push.{ASSET_KEY}
+            exec.asset_vault::get_asset
             push.{non_fungible_asset}
-            exec.asset_vault::has_non_fungible_asset
-            assert.err="vault should contain asset"
+            assert_eqw.err="vault should contain asset"
 
             dropw
         end
         "#,
+        ASSET_KEY = non_fungible_asset.vault_key(),
         non_fungible_asset = Word::from(non_fungible_asset),
     );
 
@@ -277,16 +285,11 @@ async fn mint_non_fungible_asset_fails_on_non_faucet_account() -> anyhow::Result
 
 #[tokio::test]
 async fn test_burn_fungible_asset_succeeds() -> anyhow::Result<()> {
-    let tx_context = {
-        let account = Account::mock_fungible_faucet(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1);
-        let note = create_public_p2any_note(
-            ACCOUNT_ID_SENDER.try_into().unwrap(),
-            [FungibleAsset::new(account.id(), 100u64).unwrap().into()],
-        );
-        TransactionContextBuilder::new(account).extend_input_notes(vec![note]).build()?
-    };
-
-    let faucet_id = tx_context.account().id();
+    let account = Account::mock_fungible_faucet(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1);
+    let asset = FungibleAsset::new(account.id(), 100u64).unwrap().into();
+    let note = create_public_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), [asset]);
+    let tx_context =
+        TransactionContextBuilder::new(account).extend_input_notes(vec![note]).build()?;
 
     let code = format!(
         r#"
@@ -300,25 +303,30 @@ async fn test_burn_fungible_asset_succeeds() -> anyhow::Result<()> {
             exec.prologue::prepare_transaction
 
             # burn asset
-            push.{FUNGIBLE_ASSET_AMOUNT} push.0 push.{suffix} push.{prefix}
+            push.{FUNGIBLE_ASSET}
             call.mock_faucet::burn
 
             # assert the correct asset is returned
-            push.{FUNGIBLE_ASSET_AMOUNT} push.0 push.{suffix} push.{prefix}
+            push.{FUNGIBLE_ASSET}
             assert_eqw.err="burnt asset does not match expected asset"
 
             # assert the input vault has been updated
             exec.memory::get_input_vault_root_ptr
 
-            push.{suffix} push.{prefix}
-            exec.asset_vault::get_balance
+            push.{ASSET_KEY}
+            exec.asset_vault::get_asset
+            # => [ASSET]
+
+            # extract balance from asset
+            drop drop drop
+            # => [balance]
 
             push.{final_input_vault_asset_amount}
             assert_eq.err="vault balance does not match expected balance"
         end
         "#,
-        prefix = faucet_id.prefix().as_felt(),
-        suffix = faucet_id.suffix(),
+        FUNGIBLE_ASSET = Word::from(asset),
+        ASSET_KEY = asset.vault_key(),
         final_input_vault_asset_amount = CONSUMED_ASSET_1_AMOUNT - FUNGIBLE_ASSET_AMOUNT,
     );
 
@@ -425,7 +433,6 @@ async fn test_burn_non_fungible_asset_succeeds() -> anyhow::Result<()> {
     let tx_context =
         TransactionContextBuilder::with_non_fungible_faucet(NonFungibleAsset::mock_issuer().into())
             .build()?;
-
     let non_fungible_asset_burnt = NonFungibleAsset::mock(&NON_FUNGIBLE_ASSET_DATA_2);
 
     let code = format!(
@@ -445,9 +452,10 @@ async fn test_burn_non_fungible_asset_succeeds() -> anyhow::Result<()> {
 
             # check that the non-fungible asset is presented in the input vault
             exec.memory::get_input_vault_root_ptr
+            push.{ASSET_KEY}
+            exec.asset_vault::get_asset
             push.{non_fungible_asset}
-            exec.asset_vault::has_non_fungible_asset
-            assert.err="input vault should contain the asset"
+            assert_eqw.err="input vault should contain the asset"
 
             # burn the non-fungible asset
             push.{non_fungible_asset}
@@ -459,13 +467,15 @@ async fn test_burn_non_fungible_asset_succeeds() -> anyhow::Result<()> {
 
             # assert the input vault has been updated and does not have the burnt asset
             exec.memory::get_input_vault_root_ptr
-            push.{non_fungible_asset}
-            exec.asset_vault::has_non_fungible_asset
-            not assert.err="input vault should not contain burned asset"
+            push.{ASSET_KEY}
+            exec.asset_vault::get_asset
+            # the returned word should be empty, indicating the asset is absent
+            padw assert_eqw.err="input vault should not contain burned asset"
 
             dropw
         end
         "#,
+        ASSET_KEY = non_fungible_asset_burnt.vault_key(),
         non_fungible_asset = Word::from(non_fungible_asset_burnt),
     );